requirements.txt
Django>=3.2.15,<3.2.20
djangorestframework>=3.13.1,<3.14
djangorestframework-simplejwt>=5.2.0,<5.3.0
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.23.1,<0.24
Pillow>=8.2.0,<8.3.0
uwsgi>=2.0.19,<2.1
